requirements.txt
pip==20.2.*
bump2version==0.5.11
wheel==0.33.6
watchdog==0.9.0
flake8==3.7.8
tox==3.14.0
coverage==4.5.4
Sphinx==1.8.5
twine==1.14.0

pytest==4.6.5
pytest-runner==5.1

numpy==1.18.*
pandas==1.0.*
scikit-learn==0.22.*

azureml-sdk
azure-storage-blob==12.3.0
joblib==0.14.1
requests==2.23.0
